requirements.txt
pyrogram
telegram
python-telegram-bot
requests
python-dotenv
